+++ seung-jae/src/main/java/chap2/Ex2.java
@@ -1,4 +1,4 @@
-package ex;
+package chap2;
 
 public class Ex2 {
